requirements.txt
# =============================================================================
# PRODUCTION Requirements — Agent V2 Voice AI
# =============================================================================
# Curated for production deployment. No test dependencies.
# Pin versions for reproducible builds.

# =============================================================================
# LiveKit Agent Framework (v1.3.x for AgentSession & function_tool)
# =============================================================================
livekit-agents[silero,turn-detector]==1.3.10
livekit-plugins-openai>=0.10.0
livekit-plugins-deepgram>=0.7.0
livekit-plugins-assemblyai>=0.3.0
livekit-plugins-cartesia>=0.4.0
livekit-plugins-noise-cancellation>=0.3.0

# =============================================================================
# OpenAI (GPT-4o-mini & Embeddings)
# =============================================================================
openai>=1.40.0

# =============================================================================
# Google Calendar API
# =============================================================================
google-api-python-client>=2.140.0
google-auth>=2.32.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# =============================================================================
# Contact & Date Utilities
# =============================================================================
phonenumbers>=8.13.40
email-validator>=2.2.0
python-dateutil>=2.9.0

# =============================================================================
# Database (Supabase)
# =============================================================================
supabase>=2.10.0

# =============================================================================
# Environment & Config
# =============================================================================
python-dotenv>=1.0.1

# =============================================================================
# Timezone Support (CRITICAL for Windows deployments)
# =============================================================================
tzdata>=2024.1

# =============================================================================
# Async HTTP (for webhooks/callbacks)
# =============================================================================
aiohttp>=3.10.0